refactor: move L1Event to btcio
