+++ crates/btcio/src/reader/handler.rs
@@ -1,5 +1,4 @@
 use bitcoin::{consensus::serialize, hashes::Hash, Block};
-use strata_l1tx::messages::{BlockData, L1Event};
 use strata_primitives::{
     buf::Buf32,
     l1::{
@@ -10,7 +9,10 @@ use strata_primitives::{
 use strata_state::sync_event::{EventSubmitter, SyncEvent};
 use tracing::*;
 
-use super::query::ReaderContext;
+use super::{
+    event::{BlockData, L1Event},
+    query::ReaderContext,
+};
 use crate::rpc::traits::ReaderRpc;
 
 pub(crate) async fn handle_bitcoin_event<R: ReaderRpc>(
