+++ crates/btcio/src/reader/mod.rs
@@ -1,3 +1,4 @@
+mod event;
 mod handler;
 pub mod query;
 mod state;
